+++ o11y-kit/o11y-kit-spring-boot-autoconfigure/src/main/java/io/o11y/kit/autoconfigure/ObservationWebFluxAutoConfiguration.java
@@ -5,6 +5,7 @@ import io.o11y.kit.spring.webflux.ClientObservationHandler;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -14,6 +15,7 @@ import org.springframework.web.reactive.function.client.WebClient;
  * for WebClient-based outbound HTTP callout metrics.
  *
  * <p>Activates only in a reactive web application context.
+ * Can be disabled by setting {@code o11y.kit.client.enabled=false}.
  * Users can apply this handler to individual WebClient instances via:
  * <pre>{@code
  *   webClientBuilder.filter(clientObservationHandler);
@@ -24,6 +26,7 @@ import org.springframework.web.reactive.function.client.WebClient;
 @AutoConfiguration
 @ConditionalOnClass(WebClient.class)
 @ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.REACTIVE)
+@ConditionalOnProperty(prefix = "o11y.kit.client", name = "enabled", matchIfMissing = true)
 public class ObservationWebFluxAutoConfiguration {
 
     @Bean